New feature: User-made projects!
Users can now create project directories with a name of their choice!

Upon starting the program, Lemon Notes will check for a "Projects"
directory. If no such folder exists, or the folder is empty, the user
will be prompted to enter a name for their first project. If no valid
name is entered, a default directory will be built.

Once at least one project exists, the user will be sent to the main
editor window. If they wish to create a new project, they may do so with
the "New Project" button.

**Note: The regex for the subject name and project name now allows for
underscores and spaces in addition to alphanumeric characters.