Implemented bold and italics
Added the ability to format text to be either bold or italicised. (Note:
as of this build, it isn't possible to both italicise and bold a piece
of text)

To make text bold, do the following: <b>This text will be bold</b>
For italics: <i>Italicised text!</i>

Side note: This was just a quick and dirty implementation to test the
parsing of tags, and does not take advantage of the mode classes. I will
fix this in the next commit.